[skip ci] feat: two hard things in software
