--- src/lib/fees/parameters/swapper.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_FEE_BPS = 55 // basis points
