+++ src/pages/ThorChainLP/components/AddLiquidity/AddLiquidityInput.tsx
@@ -62,8 +62,8 @@ import { useToggle } from '@/hooks/useToggle/useToggle'
 import { useWallet } from '@/hooks/useWallet/useWallet'
 import { walletSupportsChain } from '@/hooks/useWalletSupportsChain/useWalletSupportsChain'
 import { bn, bnOrZero, convertPrecision } from '@/lib/bignumber/bignumber'
-import { calculateFeeUsd } from '@/lib/fees/model'
-import { DEFAULT_FEE_BPS } from '@/lib/fees/parameters/swapper'
+import { DEFAULT_FEE_BPS } from '@/lib/fees/constant'
+import { calculateFeeUsd } from '@/lib/fees/utils'
 import { fromBaseUnit, toBaseUnit } from '@/lib/math'
 import { getMixPanel } from '@/lib/mixpanel/mixPanelSingleton'
 import { MixPanelEvent } from '@/lib/mixpanel/types'
