+++ src/state/slices/tradeQuoteSlice/selectors.ts
@@ -27,7 +27,7 @@ import { SWAPPER_USER_ERRORS } from './constants'
 import type { ActiveQuoteMeta } from './types'
 
 import { bn, bnOrZero } from '@/lib/bignumber/bignumber'
-import { calculateFeeUsd } from '@/lib/fees/model'
+import { calculateFeeUsd } from '@/lib/fees/utils'
 import { fromBaseUnit } from '@/lib/math'
 import { selectCalculatedFeeUsd } from '@/state/apis/snapshot/selectors'
 import { validateQuoteRequest } from '@/state/apis/swapper/helpers/validateQuoteRequest'
